Relocates the Component and Filter extension-type interface code files
to under "extensions"
