+++ src/Umbraco.Web.UI.Client/src/packages/ufm/components/manifests.ts
@@ -1,4 +1,4 @@
-import type { ManifestUfmComponent } from '../ufm-component.extension.js';
+import type { ManifestUfmComponent } from '../extensions/ufm-component.extension.js';
 
 export const manifests: Array<ManifestUfmComponent> = [
 	{
