+++ src/Umbraco.Web.UI.Client/src/packages/ufm/extensions/types.ts
@@ -1,2 +1,4 @@
 export type * from './marked-extension.extension.js';
+export type * from './ufm-filter.extension.js';
+export type * from './ufm-component.extension.js';
 export type { UmbUfmMarkedExtensionApi } from './ufm-marked-extension.api.js';
