+++ src/Umbraco.Web.UI.Client/src/packages/ufm/filters/base.filter.ts
@@ -1,4 +1,4 @@
-import type { UmbUfmFilterApi } from '../ufm-filter.extension.js';
+import type { UmbUfmFilterApi } from '../extensions/ufm-filter.extension.js';
 
 export abstract class UmbUfmFilterBase implements UmbUfmFilterApi {
 	abstract filter(...args: Array<unknown>): string | undefined | null;
