+++ src/Umbraco.Web.UI.Client/src/packages/ufm/types.ts
@@ -1,2 +1,1 @@
-export type * from './ufm-filter.extension.js';
-export type * from './ufm-component.extension.js';
+export type * from './extensions/types.js';
